requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
python-multipart==0.0.9
pandas
rapidfuzz==3.6.1
sentence-transformers==2.3.1
faiss-cpu
groq==0.4.2
requests==2.31.0
numpy>=2.0,<3
python-dotenv==1.0.0
openpyxl>=3.1.5
pymongo>=4.6.0